requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
redis==4.5.5
python-dotenv==1.0.0